pytest-allure-selenium-po-demo-master/requirements.txt
allure_python_commons==2.10.0
PyAutoGUI==0.9.53
pytest==7.1.3
pywin32==304
PyYAML==6.0
requests==2.26.0
selenium==4.7.2
webdriver_manager==3.8.3
win32gui==221.6
